+++ zbark/zap_logger.go
@@ -16,7 +16,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package zapify
+package zbark
 
 import (
 	"sort"
